Day 22 parts one and two
Getting the correct answer for the example is far from sufficient to get the correct answer with my input. Firstly being in the same position more than once can be valid with different equipment. That is getting there by a different path which may be shorter and involve different Tool changes. Also the function calculating the remaining cost for A* is needs to be exact for the last step. Underestimation can yield an incorrect result.