Solution general to parts one and two
Should be able to solve for any input assuming that all input has a similar sequence of instructions (not values or choice of registers. I must have got close to a solution in 2018 I found a list of the first ten values that cause termination. Again Megatrends to the rescue one post suggested storing a set of all values that cause termination.
The last value that causes termination before the any repeats occur is the one which causes the most instructions to be executed. The termination values are generated by what appears to be a congruence pseudo random number  generator with a sequence length around 10000 for my input. The generator produces a small subset of the possible values 24 bit. The interpreter takes just over 1000s to find all the values that cause termination. In the problem description referring to the largest and smallest values is a red herring. Effectively asking for the value that causes termination after the least instructions (part 1) or the most (part 2) will lead to the same answer. without thoughts of binary searches ect!