Day 23 part one and flawed but working part two
Note this is not a general solution and relies in the input being flawed. It "works" by converting the three dimensional problem into one dimension by taking the Manhattan distance to the nanobot and counting the number of line segments that overlap based on that distance plus and minus the range. The line segments are limited to starting at 0. This is a shameless translation of one of the Megatrend solutions.
 In the Manhattan space, the face of the octahedron nearest to the origin is a constant distance from the origin. The distance where Near_End is True in my solution, thus even if there is a unique point where the in-range count is maximum it must lie on one such face. Unless the input data is constructed with a fairly even distribution of nanobots this flawed strategy stands a reasonable chance of working!
I had a solution based on scaling which got the correct answer for the example but failed for with my input; however it demonstrated that something like 98% on the nanobots were in range of the point I found.
Effectively the nanobots are clumped in one direction relative to the
      -- origin. This also contributes to flawed solutions like this working